+++ server/src/main/java/org/apache/druid/client/indexing/ClientCompactionTaskTransformSpec.java
@@ -21,9 +21,12 @@ package org.apache.druid.client.indexing;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.druid.query.filter.DimFilter;
 
 import javax.annotation.Nullable;
+import java.util.Map;
 import java.util.Objects;
 
 /**
@@ -50,6 +53,14 @@ public class ClientCompactionTaskTransformSpec
   public DimFilter getFilter()
   {
     return filter;
+  }
+
+  public Map<String, Object> asMap(ObjectMapper objectMapper)
+  {
+    return objectMapper.convertValue(
+        this,
+        new TypeReference<Map<String, Object>>() {}
+    );
   }
 
   @Override
